requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
aiosqlite
alembic
passlib
bcrypt